# internal/model/dto/paginatedUsersDto.go
package dto

type PaginatedUsersDto struct {
	Data       []UserDto `json:"data"`
	Total      int       `json:"total"`
	Page       int       `json:"page"`
	PageSize   int       `json:"page_size"`
	TotalPages int       `json:"total_pages"`
}